refact: put productions in separate modules
