+++ src/production/__init__.py
@@ -1,0 +1,11 @@
+from .model import Production
+from .production_1 import P1
+from .production_2 import P2
+
+
+__all__ = (
+    Production,
+    P1,
+    P2
+)
+
